Bug fix: special characters in file names
structView: Section cells in the KBBrowser get a global dynamic variable,
in which the state of the openerView is stored. This variable depends on
the file name of the notebook. The automatically generated symbol was
problematic when the file name contained special symbols, e.g. "-". Now, the
file name is hashed so that this does not give troubles anymore.