Parsing expressions with values
freshSymbols uses ToExpression in order to generate a symbol. If, e.g. in
a computation, we use knowledge x==1, then the generated symbol x$TM has a
value and ToExpression evaluates. Therefore, we wrap the fresh symbol in
Hold (using ToExpression[..., Hold]) and process the resulting expression
with the expectation that all symbols are wrapped in Hold. (-> markVariables, etc.)